requirements.txt
langchain==1.0.5  
langchain-community==0.4.1  
langchain-text-splitters==1.0.0  
langchain_huggingface
langchain-classic
langchain-core
langchain_groq
faiss-cpu
pypdf
huggingface_hub
flask
python-dotenv
sentence-transformers
ollama==0.13.0
